Update README
Installing libusb-1.0-0-dev was what finally resolved a `not found LIBUSB INSTALL DIR` error when running the cmake part of installing libftdi1-1.1.

Did not seem to need libconfuse...I am not that handy with Linux but `dpkg --get-selections | grep libcon` shows no results.  It's not listed in the libftdi1-1.1 README either.  I'll leave it in there just in case.